+++ JSOAuction.Domain.Entities/PlayerRegister/PlayerRegister.cs
@@ -49,7 +49,7 @@ namespace JSOAuction.Domain.Entities.PlayerRegister
         public string OccupationTypes { get; set; }
         public string TShirtNumber { get; set; }
         public string TShirtName { get; set; }
-        public double? DonationAmount { get; set; }
+        public decimal? DonationAmount { get; set; }
         public int TShirtSizeId { get; set; }
 
     }
